# lib/openapi/generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RealEST Property Marketplace API",
    "description": "Nigerian property marketplace API with ML validation, geospatial search, and intelligent matching",
    "version": "2.0.0",
    "contact": {
      "name": "RealEST Support",
      "email": "support@realest.ng"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://realest.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT Auth token"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/emails/campaigns/{id}/send": {
      "post": {
        "method": "post",
        "summary": "Send email campaign",
        "description": "Execute a draft or scheduled campaign send using Resend broadcast or batch delivery.",
        "tags": [
          "Admin - Emails"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign sent successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Campaign not found"
          },
          "409": {
            "description": "Campaign cannot be sent in its current status"
          },
          "500": {
            "description": "Campaign send failed"
          }
        },
        "operationId": "postAdminEmailsCampaignsIdSend"
      }
    },
...
    "/api/admin/users/{id}": {
      "get": {
        "method": "get",
        "summary": "Get user details",
        "description": "Admin endpoint to fetch a single user profile, stats, properties and recent inquiries.",
        "tags": [
          "Admin - Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "operationId": "getAdminUsersId"
      }
    },
...
  },
...
}